requirements.txt
# Core LangChain
langchain==0.3.25
langchain-community==0.3.23
langchain-openai==0.3.16

# Vector Search
faiss-cpu==1.10.0
rank_bm25==0.2.2

# Agent Orchestration
langgraph==0.3.21

# UI
gradio==5.25.0

# Environment
python-dotenv==1.1.0

# Embeddings (multilingual-e5-small)
sentence-transformers==4.1.0
